requirements.txt
aiogram
python-dotenv
openai-whisper
ffmpeg-python
